Compensate for JAVA_TOOL_OPTIONS environment variable
If this environment variable is set then we won't get valid JSON from the validator - it'll be prefixed by a message from Java.
We filter the environment for the Java process to exclude this and pass these arguments on the command line instead.

Also, improve error message for invalid JSON from Java.